fix requesting  location permission multiple times
