+++ lib/accommodation_views/home.dart
