+++ lib/accommodation_views/hotels_list.dart
@@ -3,30 +3,22 @@ import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:hue_t/animation/show_up.dart';
-import 'animation/show_right.dart';
-import 'colors.dart' as colors;
+import '../animation/show_right.dart';
+import '../colors.dart' as colors;
 import 'hotel_detail.dart';
-import 'fake_data.dart' as faker;
+import '../fake_data.dart' as faker;
+import 'package:hue_t/get_user_location.dart' as userLocation;
 
-class ResortsPage extends StatefulWidget {
-  const ResortsPage({Key? key}) : super(key: key);
+class HotelsPage extends StatefulWidget {
+  const HotelsPage({Key? key}) : super(key: key);
 
   @override
-  State<ResortsPage> createState() => _ResortsPageState();
+  State<HotelsPage> createState() => _HotelsPageState();
 }
 
 bool isRecommendationHotel = true;
 
-class _ResortsPageState extends State<ResortsPage> {
-  Future<Position> getUserCurrentLocation() async {
-    await Geolocator.requestPermission()
-        .then((value) {})
-        .onError((error, stackTrace) async {
-      await Geolocator.requestPermission();
-      print("ERROR" + error.toString());
-    });
-    return await Geolocator.getCurrentPosition();
-  }
+class _HotelsPageState extends State<HotelsPage> {
 
   Future<void> distanceCaculating(Position value) async {
     for (int i = 0; i < faker.listHotels.length; i++) {
@@ -45,7 +37,7 @@ class _ResortsPageState extends State<ResortsPage> {
   @override
   Widget build(BuildContext context) {
     if (isLoading) {
-      getUserCurrentLocation().then((value) async {
+      userLocation.getUserCurrentLocation().then((value) async {
         print(value.latitude.toString() + " " + value.longitude.toString());
         await distanceCaculating(value);
         setState(() {
@@ -67,12 +59,12 @@ class _ResortsPageState extends State<ResortsPage> {
       );
     }
     return Scaffold(
-      body: Stack(
-        children: [
-          contentBlock(context),
-          backButton(context)
-        ],
-      ),
+        body: Stack(
+          children: [
+            contentBlock(context),
+            backButton(context)
+          ],
+        ),
     );
   }
 
@@ -209,11 +201,11 @@ class _ResortsPageState extends State<ResortsPage> {
       child: Container(
         height: MediaQuery.of(context).size.height/4,
         decoration: BoxDecoration(
-            image: DecorationImage(
-                image: NetworkImage("https://cdn4.tropicalsky.co.uk/images/1800x600/indochine-palace-main-image.jpg"),
-                fit: BoxFit.cover,
-                alignment: Alignment.topCenter
-            )
+          image: DecorationImage(
+            image: NetworkImage("https://cdn4.tropicalsky.co.uk/images/1800x600/indochine-palace-main-image.jpg"),
+            fit: BoxFit.cover,
+            alignment: Alignment.topCenter
+          )
         ),
       ),
     );
@@ -226,8 +218,8 @@ class _ResortsPageState extends State<ResortsPage> {
         child: Center(
           child: Column(
             children: [
-              Text("Resorts", style: GoogleFonts.poppins(fontSize: 25, fontWeight: FontWeight.bold, color: Colors.black), textAlign: TextAlign.center,),
-              Text("Best-rated resort in the last month", style: GoogleFonts.poppins(fontSize: 15, fontWeight: FontWeight.w300), textAlign: TextAlign.center),
+              Text("Hot Hotels", style: GoogleFonts.poppins(fontSize: 25, fontWeight: FontWeight.bold, color: Colors.black), textAlign: TextAlign.center,),
+              Text("Best-rated hotels in the last month", style: GoogleFonts.poppins(fontSize: 15, fontWeight: FontWeight.w300), textAlign: TextAlign.center),
             ],
           ),
         ),
